requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
attrs==24.2.0
blinker==1.9.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
discord.py==2.4.0
Flask==3.1.0
frozenlist==1.5.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
mutagen==1.47.0
propcache==0.2.0
pycparser==2.22
pycryptodomex==3.21.0
PyNaCl==1.5.0
requests==2.32.3
urllib3==2.2.3
websockets==13.1
Werkzeug==3.1.3
yarl==1.18.0
youtube-dl==2021.12.17
yt-dlp==2024.11.18
